backend/requirements.txt
# Dependencias principales del backend FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Autodesk Platform Services SDK
aps-sdk-python==1.0.1
requests==2.31.0
httpx==0.25.2

# OAuth 2.0 y autenticación adicional
authlib==1.2.1
itsdangerous==2.1.2

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Cache y sesiones
redis==5.0.1
celery==5.3.4

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Middleware y CORS
python-cors==1.7.0

# Logging y monitoreo
structlog==23.2.0
sentry-sdk==1.38.0

# Utilidades
python-dotenv==1.0.0
pydantic-extra-types==2.2.0
email-validator==2.1.0

# File processing and validation
python-magic==0.4.27
Pillow==10.1.0

# Background tasks
celery[redis]==5.3.4

# Additional security
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.0.0
pytest-mock==3.12.0
httpx==0.25.2

# Desarrollo
black==23.11.0
flake8==6.1.0
mypy==1.7.1